+++ Magic/src/main/java/com/elmakers/mine/bukkit/action/builtin/RepeatAction.java
@@ -37,7 +37,7 @@ public class RepeatAction extends CompoundAction
     @Override
     public SpellResult step(CastContext context) {
         // special case for variable repeat counts
-        if (count == 0) return SpellResult.NO_ACTION;
+        if (count == 0 && !infinite) return SpellResult.NO_ACTION;
         return startActions();
     }