+++ Magic/CHANGELOG.md
@@ -3,6 +3,7 @@
 # 7.6.16
 
  - Fix some hangs that could happen when using an infinite Repeat action
+ - Fix using `repeat: infinite`
 
 # 7.6.15